backend/services/gamification-service/requirements.txt
fastapi>=0.104.0
pydantic>=2.0.0
boto3>=1.28.0
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
uvicorn[standard]>=0.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
moto>=4.2.0
httpx>=0.24.0
